code optimization src/Main.java

+++ src/Main.java
@@ -2,9 +2,9 @@ public class Main {
     public static void main(String[] args) {
 
         int ticket = 144_544;
-        float condition = 20F;
+        byte condition = 20;
 
-        int bonusMiles = (int) (ticket / condition);
+        int bonusMiles = ticket / condition;
 
         System.out.println(bonusMiles);
     }
